pyproject.toml
[tool.poetry]
name = "self-rag-chatbot"
version = "0.1.0"
description = ""
authors = ["Brais Maneiro Sánchez <brais.maneiro@netexlearning.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.3.3"
langgraph = "^0.2.38"
beautifulsoup4 = "^4.12.3"
langchain-community = "^0.3.2"
tavily-python = "^0.5.0"
langchain-chroma = "^0.1.4"
python-dotenv = "^1.0.1"
pytest = "^8.3.3"
langchain-text-splitters = "^0.3.0"
langchain-openai = "^0.2.3"
langchain-exa = "^0.2.0"
langfuse = "^2.53.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
